Restore MG2 compatibility in onQueueComplete
Aborting the current handler and short-circuiting to exception handler ignores onQueueComplete. This broke compatibility with MG2.  Regardless of use case, onQueueComplete should fire even when an exception occurs.  Regardless of why queue processing is done, the event should be fired.  

A common use case would be to share presentation configuration information that is not required until all other processing is complete, e.g.:

{{{
<event-handler name="ModelGlue.onQueueComplete">
	<broadcasts>
		<message name="needSiteTheme" />
	</broadcasts>
</event-handler>
}}}

Why put it in onQueueComplete?  Consider that some other processing might determine which theme to display in a themeable system.